added DB working code for guides
