+++ app/(web)/matterport.html
@@ -1,0 +1,74 @@
+<!DOCTYPE html>
+<html>
+<head>
+  <title>Matterport Viewer</title>
+  <meta charset="UTF-8">
+  <style>
+    body, html, #showcase {
+      margin: 0;
+      padding: 0;
+      height: 100%;
+      width: 100%;
+    }
+    #capture-button {
+      position: absolute;
+      top: 10px;
+      left: 10px;
+      z-index: 1000;
+      padding: 10px 20px;
+      font-size: 16px;
+    }
+  </style>
+  <!-- Include the Matterport SDK script -->
+  <script src="https://static.matterport.com/showcase-sdk/latest.js"></script>
+</head>
+<body>
+  <div id="showcase"></div>
+  <button id="capture-button">Capture Snapshot</button>
+  <script>
+    // Replace with your Matterport SDK key
+    const SDK_KEY = 'YOUR_SDK_KEY';
+
+    // Replace with your Matterport space ID
+    const SPACE_ID = 'YOUR_SPACE_ID';
+
+    // Initialize the Matterport SDK
+    window.MP_SDK.connect(
+      document.getElementById('showcase'),
+      SDK_KEY,
+      SPACE_ID,
+      {
+        // SDK options
+      }
+    ).then(function (sdk) {
+      console.log('Matterport SDK connected');
+
+      // Function to capture a snapshot
+      function captureSnapshot() {
+        sdk.Renderer.takeScreenShot().then(function (screenshotURI) {
+          // Send the screenshot data to the React Native app
+          window.ReactNativeWebView.postMessage(JSON.stringify({
+            type: 'snapshot',
+            data: screenshotURI
+          }));
+        }).catch(function (error) {
+          console.error('Error taking screenshot:', error);
+          window.ReactNativeWebView.postMessage(JSON.stringify({
+            type: 'error',
+            message: error.message
+          }));
+        });
+      }
+
+      // Add event listener to the capture button
+      document.getElementById('capture-button').addEventListener('click', captureSnapshot);
+    }).catch(function (error) {
+      console.error('Error connecting to Matterport SDK:', error);
+      window.ReactNativeWebView.postMessage(JSON.stringify({
+        type: 'error',
+        message: error.message
+      }));
+    });
+  </script>
+</body>
+</html>
