+++ app/_layout.tsx
@@ -1,16 +1,20 @@
-import { ClerkProvider } from '@clerk/clerk-expo';
-import { DarkTheme, DefaultTheme, ThemeProvider } from '@react-navigation/native';
-import { useFonts } from 'expo-font';
-import { Stack } from 'expo-router';
-import * as SplashScreen from 'expo-splash-screen';
-import { useEffect } from 'react';
-import React from 'react';
-import 'react-native-reanimated';
+import { ClerkProvider } from "@clerk/clerk-expo";
+import {
+  DarkTheme,
+  DefaultTheme,
+  ThemeProvider,
+} from "@react-navigation/native";
+import { useFonts } from "expo-font";
+import { Stack } from "expo-router";
+import * as SplashScreen from "expo-splash-screen";
+import { useEffect } from "react";
+import React from "react";
+import "react-native-reanimated";
 
 // Prevent the splash screen from auto-hiding before asset loading is complete.
 SplashScreen.preventAutoHideAsync();
 
-const publishableKey = process.env.EXPO_PUBLIC_CLERK_PUBLISHABLE_KEY!
+const publishableKey = process.env.EXPO_PUBLIC_CLERK_PUBLISHABLE_KEY!;
 
 export default function RootLayout() {
   const [loaded] = useFonts({
@@ -21,7 +25,7 @@ export default function RootLayout() {
     "Jakarta-Medium": require("../assets/fonts/PlusJakartaSans-Medium.ttf"),
     "Jakarta-Regular": require("../assets/fonts/PlusJakartaSans-Regular.ttf"),
     "Jakarta-SemiBold": require("../assets/fonts/PlusJakartaSans-SemiBold.ttf"),
-});
+  });
 
   useEffect(() => {
     if (loaded) {
@@ -34,19 +38,18 @@ export default function RootLayout() {
   }
 
   return (
-    // the naming here does not matter because the first publishableKey is the PROPERTY name. and the second one is the variable we are assigning it to. 
+    // the naming here does not matter because the first publishableKey is the PROPERTY name. and the second one is the variable we are assigning it to.
     <ClerkProvider publishableKey={publishableKey}>
       <Stack>
         <Stack.Screen name="index" options={{ headerShown: false }} />
         <Stack.Screen name="(root)" options={{ headerShown: false }} />
         <Stack.Screen name="(auth)" options={{ headerShown: false }} />
+        <Stack.Screen name="(guides)" options={{ headerShown: false }} />
         <Stack.Screen name="+not-found" />
-        <Stack.Screen name = "(app)" options={{ headerShown: false }} />
-        <Stack.Screen name = "log-in" options={{ headerShown: false }} />
+        <Stack.Screen name="(app)" options={{ headerShown: false }} />
+        <Stack.Screen name="log-in" options={{ headerShown: false }} />
         <Stack.Screen name="(realtor)" options={{ headerShown: false }} />
-
       </Stack>
-    </ClerkProvider> 
+    </ClerkProvider>
   );
 }
-
