+++ types/type.d.ts
@@ -1,6 +1,7 @@
 import {TextInputProps, TouchableOpacityProps} from "react-native";
 
 declare interface Question {
+    id: number;
     question_text: string;
     options: string[];
     correct_answer: string;
